requirements.txt
pyyaml
pathlib
scikit-learn
numpy
scipy
astropy
pytest
pathlib
